Refactor: ItemPrototypeFactory is wrongly located at internal package
