+++ src/main/java/io/ib67/astralflow/api/item/machine/MachineCategory.java
@@ -23,7 +23,7 @@ package io.ib67.astralflow.api.item.machine;
 
 import io.ib67.astralflow.item.builder.ItemCategory;
 import io.ib67.astralflow.item.builder.ItemPrototype;
-import io.ib67.astralflow.item.internal.factory.ItemPrototypeFactory;
+import io.ib67.astralflow.item.factory.ItemPrototypeFactory;
 import org.jetbrains.annotations.ApiStatus;
 
 /**
