+++ src/main/java/io/ib67/astralflow/api/item/weapon/WeaponCategory.java
@@ -23,7 +23,7 @@ package io.ib67.astralflow.api.item.weapon;
 
 import io.ib67.astralflow.item.builder.ItemCategory;
 import io.ib67.astralflow.item.builder.ItemPrototype;
-import io.ib67.astralflow.item.internal.factory.ItemPrototypeFactory;
+import io.ib67.astralflow.item.factory.ItemPrototypeFactory;
 
 public final class WeaponCategory implements ItemCategory<WeaponBase> {
     public static final WeaponCategory INSTANCE = new WeaponCategory();
