+++ src/main/java/io/ib67/astralflow/item/recipe/choices/AstralItemChoice.java
@@ -25,7 +25,7 @@ import io.ib67.astralflow.AstralFlow;
 import io.ib67.astralflow.internal.item.state.InternalItemState;
 import io.ib67.astralflow.item.ItemKey;
 import io.ib67.astralflow.item.StateScope;
-import io.ib67.astralflow.item.internal.factory.ItemPrototypeFactory;
+import io.ib67.astralflow.item.factory.ItemPrototypeFactory;
 import io.ib67.astralflow.item.recipe.IngredientChoice;
 import io.ib67.util.Lazy;
 import lombok.Getter;
