+++ src/main/java/io/ib67/astralflow/manager/impl/ItemRegistryImpl.java
@@ -26,7 +26,7 @@ import io.ib67.astralflow.item.AstralItem;
 import io.ib67.astralflow.item.ItemKey;
 import io.ib67.astralflow.item.ItemState;
 import io.ib67.astralflow.item.StateScope;
-import io.ib67.astralflow.item.internal.factory.ItemPrototypeFactory;
+import io.ib67.astralflow.item.factory.ItemPrototypeFactory;
 import io.ib67.astralflow.item.internal.serialization.ItemStorageType;
 import io.ib67.astralflow.item.internal.tag.ItemStateTag;
 import io.ib67.astralflow.item.oredict.IOreDict;
